+++ common/src/main/java/com/dudko/blazinghot/content/item/BlazeArrowItem.java
@@ -3,7 +3,6 @@ package com.dudko.blazinghot.content.item;
 import com.dudko.blazinghot.content.entity.BlazeArrowEntity;
 import com.dudko.blazinghot.registry.BlazingEntities;
 import com.dudko.blazinghot.registry.BlazingItems;
-import com.simibubi.create.content.contraptions.ContraptionWorld;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.projectile.AbstractArrow;
 import net.minecraft.world.item.ArrowItem;
@@ -18,11 +17,6 @@ public class BlazeArrowItem extends ArrowItem {
 
     @Override
     public @NotNull AbstractArrow createArrow(@NotNull Level level, @NotNull ItemStack stack, @NotNull LivingEntity shooter) {
-        BlazeArrowEntity
-                arrow =
-                new BlazeArrowEntity(BlazingEntities.BLAZE_ARROW.get(), shooter, level, BlazingItems.BLAZE_ARROW.get());
-        if (level.dimension() == ContraptionWorld.NETHER) arrow.setBaseDamage(arrow.getBaseDamage() * 1.5);
-
-        return arrow;
+        return new BlazeArrowEntity(BlazingEntities.BLAZE_ARROW.get(), shooter, level);
     }
 }