+++ common/src/main/java/com/dudko/blazinghot/mixin/SpoutBlockMixin.java
@@ -12,7 +12,7 @@ import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
 import org.spongepowered.asm.mixin.injection.callback.CallbackInfo;
 
-@Mixin(value = SpoutBlock.class, remap = false)
+@Mixin(value = SpoutBlock.class)
 public class SpoutBlockMixin {
 
     @Inject(method = "setPlacedBy", at = @At("TAIL"))