+++ common/src/main/java/com/dudko/blazinghot/mixin/accessor/ProjectileAccessor.java
@@ -1,0 +1,15 @@
+package com.dudko.blazinghot.mixin.accessor;
+
+import net.minecraft.world.entity.projectile.Projectile;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.gen.Accessor;
+
+import java.util.UUID;
+
+@Mixin(Projectile.class)
+public interface ProjectileAccessor {
+
+    @Accessor
+    UUID getOwnerUUID();
+
+}